♻️ Replace custom impl of Coins with cw_std Coins
