+++ packages/skip/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod coins;
 pub mod entry_point;
 pub mod error;
 pub mod ibc;
